Various changes
Update to Sections 5 and 6, incorporating Ibs changes and trying to tell an analytics storyline.

Both the .decomp and the attribution discussions would work better if we had tabular outputs in addition to the graphs.  This is also true for the tracedep and tracepre.  All of this should be reviewed by Ib for extension and verificationl.